Remove unnecessary casts.
git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@36938 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ gui/dialog/skinbrowser.c
@@ -169,7 +169,7 @@ int FillSkinList( gchar * mdir )
 {
  gchar         * str[2];
  gchar         * tmp;
- int             i;
+ size_t          i;
  glob_t          gg;
  struct stat     fs;
 
@@ -181,7 +181,7 @@ int FillSkinList( gchar * mdir )
  if ( gtkFindInCList( SkinList,str[0] ) == -1 ) gtk_clist_append( GTK_CLIST( SkinList ),str );
 
  glob( mdir,GLOB_NOSORT,NULL,&gg );
- for( i=0;i<(int)gg.gl_pathc;i++ )
+ for( i=0;i<gg.gl_pathc;i++ )
   {
    if ( !strcmp( gg.gl_pathv[i],"." ) || !strcmp( gg.gl_pathv[i],".." ) ) continue;
    if ( ( stat( gg.gl_pathv[i],&fs ) == 0 ) && S_ISDIR( fs.st_mode ) )
